requirements.txt
google-cloud-bigquery==3.17.2